# halve allocations