--- a/graphql/executor/path.go
+++ b/graphql/executor/path.go
@@ -1,14 +1,22 @@
 package executor
 
 type path struct {
-	Prev      *path
-	Component interface{}
+	Prev            *path
+	StringComponent string
+	IntComponent    int
 }
 
-func (p *path) WithComponent(component interface{}) *path {
+func (p *path) WithIntComponent(n int) *path {
 	return &path{
-		Prev:      p,
-		Component: component,
+		Prev:         p,
+		IntComponent: n,
+	}
+}
+
+func (p *path) WithStringComponent(s string) *path {
+	return &path{
+		Prev:            p,
+		StringComponent: s,
 	}
 }
 
@@ -16,5 +24,8 @@ func (p *path) Slice() []interface{} {
 	if p == nil {
 		return nil
 	}
-	return append(p.Prev.Slice(), p.Component)
+	if p.StringComponent != "" {
+		return append(p.Prev.Slice(), p.StringComponent)
+	}
+	return append(p.Prev.Slice(), p.IntComponent)
 }