## What it Does

* Eliminates excessive closure / future generation for nullable error catching
* Eliminates `interface{}` usage in path construction
* Eliminates argument map allocation for fields with no arguments
* Reduces closure allocation in future functions
* Eliminates `interface{}` allocation in string coercion

Altogether this more than halves allocations.

Before:

```
chris@mbp18 executor % go test . -run asdqwe -bench BenchmarkExecuteRequest
goos: darwin
goarch: amd64
pkg: github.com/ccbrown/api-fu/graphql/executor
BenchmarkExecuteRequest-12    	     607	   1960918 ns/op	 1415254 B/op	   42777 allocs/op
PASS
ok  	github.com/ccbrown/api-fu/graphql/executor	1.516s
go test . -run asdqwe -bench BenchmarkExecuteRequest  2.73s user 0.62s system 161% cpu 2.074 total
```

After:

```
chris@mbp18 executor % go test . -run asdqwe -bench BenchmarkExecuteRequest
goos: darwin
goarch: amd64
pkg: github.com/ccbrown/api-fu/graphql/executor
BenchmarkExecuteRequest-12    	    1017	   1175525 ns/op	  813230 B/op	   20465 allocs/op
PASS
ok  	github.com/ccbrown/api-fu/graphql/executor	1.409s
go test . -run asdqwe -bench BenchmarkExecuteRequest  2.48s user 0.63s system 157% cpu 1.973 total
```

## Steps to Test

<!-- All changes should have automated tests when feasible: -->

`go test -v ./...`

<!-- Does running this require any special setup or dependencies? -->